Use 16 bit for sequence number as we don't need more

+++ src/mcs/streaming/rtpsender.h
@@ -61,7 +61,7 @@ private:
     network::Stream::Ptr stream_;
     const std::uint32_t max_ts_packets_;
     video::SenderReport::Ptr report_;
-    uint32_t rtp_sequence_number_;
+    uint16_t rtp_sequence_number_;
     mcs::video::BufferQueue::Ptr queue_;
 };
